# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestão Jurídica",
        "description": "Sistema de Gestão Jurídica",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Listar Pedidos",
                "operationId": "pedido/index",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "order  accepts 'asc' or 'desc'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pedido": {
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Cadastrar Pedido",
                "operationId": "pedido/store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "status"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Pedido Teste"
                                    },
                                    "logradouro": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Logradouro Teste"
                                    },
                                    "numero": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "99"
                                    },
                                    "bairro": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Bairro Teste"
                                    },
                                    "cidade": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Cidade Teste"
                                    },
                                    "cep": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "99999999"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "SP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pedido/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Pesquisar Pedido",
                "operationId": "pedido/show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Editar Pedido",
                "operationId": "pedido/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "status"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Pedido Teste"
                                    },
                                    "logradouro": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Logradouro Teste"
                                    },
                                    "numero": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "99"
                                    },
                                    "bairro": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Bairro Teste"
                                    },
                                    "cidade": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Cidade Teste"
                                    },
                                    "cep": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "99999999"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "SP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Excluir Pedido",
                "operationId": "pedido/destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}